third_party/blink/web_tests/wpt_internal/webrtc/navigator-rtc.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_true('rtc' in navigator, 'rtc in navigator');
}, 'check navigator.rtc');
</script>
